requirements.txt
alembic==1.0.10
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Limiter==1.0.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
Flask-UUID==0.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.17
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
limits==1.3
Mako==1.0.10
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.0.0
mysql-connector==2.2.9
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pyrsistent==0.15.2
pytest==4.5.0
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
